requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
torch>=2.0.0

# Optional but recommended
tqdm>=4.66.0  # Progress bars for indexing
pytest>=7.0.0  # For running tests/test_core.py

